git-svn-id: svn://192.168.1.10/public/CoffeeMud@8017 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Behaviors/DoorwayGuardian.java
@@ -96,7 +96,7 @@ public class DoorwayGuardian extends StdBehavior
 		if(dirs!=null)
 		for(Enumeration<Integer> dirE=dirs.elements();dirE.hasMoreElements();)
 		{
-			int dir=dirE.nextElement();
+			int dir=dirE.nextElement().intValue();
 			if(room.getExitInDir(dir)!=null)
 			{
 				Exit[] exits={room.getExitInDir(dir),room.getReverseExit(dir)};
